+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/builder/RequestBuilder.kt
@@ -7,5 +7,8 @@ interface RequestBuilder<T> {
 
 @KordDsl
 interface AuditRequestBuilder<T> : RequestBuilder<T> {
+    /**
+     * The reason for this request, this will be displayed in the audit log.
+     */
     var reason: String?
 }