+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/builder/user/GroupDMCreateBuilder.kt
@@ -1,9 +1,11 @@
 package com.gitlab.kordlib.rest.builder.user
 
 import com.gitlab.kordlib.common.entity.Snowflake
+import com.gitlab.kordlib.rest.builder.KordDsl
 import com.gitlab.kordlib.rest.builder.RequestBuilder
 import com.gitlab.kordlib.rest.json.request.GroupDMCreateRequest
 
+@KordDsl
 class GroupDMCreateBuilder : RequestBuilder<GroupDMCreateRequest> {
 
     val tokens: MutableList<String> = mutableListOf()